Now using express.js
The "Add Exercise" in template isn't currently working. I'm guessing
that since when we click the "Add Exercise", we are going to
localhost/3000/template#. And the server doesn't know what to do with
the # at the end.

But what I don't understand is, it's still saying "Uncaught SyntaxError:
Unexpected token <" when the page initially loads.