Laid down the concept of a session.
The session page will save a set when the button is clicked.
Then it should show in the bar for that exercise the % of sets done.
In the backend, the script file "sessionBuilder" should start creating
or appending these new sets that are saved to the proper exercises.
Then, when the user is done, the button "save session" should POST the
session object to the server. Which is then saved.