Now selectively posts sets.
Didn't realize it, but before, I was only posting the 1st exercise
progress in session.pug

Encountered a problem where when I switched the button and form id into
classes, all of them would get posted any time the "save set" button is
clicked.

Tried having a loop outside of the .click(). But that ended up
redirecting the page when save set is clicked. Because I think when the
document is ready, the loop runs and ends before a user can click
anything. When it is finally clicked, the loop already ended, and
.click() is never called.

Tried changing it so that all buttons are basically the same. and a loop
inside of .click(). But this didn't work, because I still run into the
problem of identifying which button was clicked.

A google search later, I found that I can just find out which button I
have clicked in e.target.id...

Now I just get the # of the button that was clicked, and pass it into
the ajax call.

Yay!