Cleaned up a little, should do more though.
Should use the utility functions over the dirty ones.

I've decided on D3.js for the graph visualizations.
And to select and view of a day's session from the homepage, I will be
using FullCalendar.js for the date selector. It should show a preview of
a given day's session (title, if a pr has been hit or not, etc.)